+++ src/task/tick_update.cpp
@@ -34,7 +34,8 @@ void TaskManager::TickUpdate() {
       task->fsm.Receive(MsgWakeup{});
 
       // 将任务重新加入对应调度器的就绪队列
-      auto* scheduler = cpu_sched.schedulers[task->policy].get();
+      auto* scheduler =
+          cpu_sched.schedulers[static_cast<uint8_t>(task->policy)].get();
       if (scheduler) {
         scheduler->Enqueue(task);
       }
@@ -51,7 +52,8 @@ void TaskManager::TickUpdate() {
       }
 
       // 调用调度器的 OnTick，检查是否需要抢占
-      auto* scheduler = cpu_sched.schedulers[current->policy].get();
+      auto* scheduler =
+          cpu_sched.schedulers[static_cast<uint8_t>(current->policy)].get();
 
       if (scheduler) {
         // 调度器可能基于自己的策略决定是否抢占
